blog page and link fixed
