+++ src/pages/Blog.js
@@ -1,19 +1,50 @@
-import React from 'react';
-import { Link } from 'gatsby';
-// import Header from '../components/Header/Header';
-import Layout from '../components/layout';
+import React from 'react'
+import { Link } from 'gatsby'
 
-const Blog = () => (
+import Layout from '../components/layout'
+import SEO from '../components/seo'
+import {graphql, StaticQuery } from 'gatsby'
+import Post from '../components/Post'
 
-    // <SEO title="Blog" />
-    <Layout>
-    
-    
-        <h1>"hello Blog"</h1>
-       
-        
-        </Layout>
-        )
+const SecondPage = () => (
+  <Layout>
+  
+    <SEO title="Page two" />
+    <h1>Hi from the second page</h1>
+    <p>Welcome to page 2</p>
+    <StaticQuery query={indexQuery} render={data => {
+      return (
+        <div>
+        {data.allMarkdownRemark.edges.map(({ node }) => (
+          <Post title={node.frontmatter.title}
+          author={node.frontmatter.author}
+          path={node.frontmatter.path}
+          date={node.frontmatter.date}
+          body={node.excerpt}
+          />
+        ))}
+        </div>
+      )
+    }}/>
+    <Link to="/">Go back to the homepage</Link>
+  </Layout>
+)
 
-
-export default Blog;
+const indexQuery = graphql `
+query {
+  allMarkdownRemark(sort: { fields:[frontmatter___date], order: DESC }) {
+    edges {
+      node {
+        id
+        frontmatter{
+          title
+          date
+          author
+          path
+        }
+        excerpt
+      }
+    }
+  }
+}`
+export default SecondPage
